Move formatting into ide layer
