+++ crates/djls-server/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2021"
 [dependencies]
 djls-conf = { workspace = true }
 djls-db = { workspace = true }
-djls-format = { workspace = true }
 djls-ide = { workspace = true }
 djls-semantic = { workspace = true }
 djls-source = { workspace = true }
