+++ crates/djls-server/src/server.rs
@@ -1,15 +1,12 @@
 use std::future::Future;
 use std::sync::Arc;
 
-use djls_format::FormatOutcome;
 use djls_semantic::load_template_library_cache;
 use djls_semantic::refresh_external_data;
 use djls_semantic::Db as SemanticDb;
 use djls_semantic::ProjectDb;
 use djls_source::Db as SourceDb;
 use djls_source::FileKind;
-use djls_source::LineIndex;
-use djls_source::PositionEncoding;
 use djls_workspace::TextDocument;
 use tokio::sync::oneshot;
 use tokio::sync::Mutex;
@@ -533,41 +530,19 @@ impl LanguageServer for DjangoLanguageServer {
                     return Vec::new();
                 };
                 let db = session.db();
-                let settings = db.settings();
-
-                if !settings.format().enabled() {
-                    return Vec::new();
-                }
-
                 let source = file.source(db);
+
                 if *source.kind() != FileKind::Template {
                     return Vec::new();
                 }
 
-                let result = match settings.format().backend() {
-                    djls_conf::FormatBackend::Djangofmt => {
-                        djls_format::format_template(source.as_str(), file.path(db))
-                    }
-                };
-
-                match result {
-                    Ok(FormatOutcome::Changed(formatted)) => vec![ls_types::TextEdit::new(
-                        full_document_range(
-                            source.as_str(),
-                            file.line_index(db),
-                            session.client_info().position_encoding(),
-                        ),
-                        formatted,
-                    )],
-                    Ok(FormatOutcome::Unchanged | FormatOutcome::Ignored) => Vec::new(),
-                    Err(error) => {
-                        tracing::debug!(
-                            "Formatting failed for {}: {error}",
-                            file.path(db).as_str(),
-                        );
-                        Vec::new()
-                    }
-                }
+                djls_ide::format_document(
+                    source.as_str(),
+                    file.path(db),
+                    file.line_index(db),
+                    session.client_info().position_encoding(),
+                    db.settings().format(),
+                )
             })
             .await;
 
@@ -630,75 +605,5 @@ impl LanguageServer for DjangoLanguageServer {
         if settings_update.env_changed || settings_update.diagnostics_changed {
             self.republish_open_template_diagnostics().await;
         }
-    }
-}
-
-fn full_document_range(
-    source: &str,
-    line_index: &LineIndex,
-    encoding: PositionEncoding,
-) -> ls_types::Range {
-    ls_types::Range {
-        start: ls_types::Position::new(0, 0),
-        end: document_end_position(source, line_index, encoding),
-    }
-}
-
-fn document_end_position(
-    source: &str,
-    line_index: &LineIndex,
-    encoding: PositionEncoding,
-) -> ls_types::Position {
-    let line = u32::try_from(line_index.lines().len().saturating_sub(1)).unwrap_or_default();
-    let line_start = line_index.lines().last().copied().unwrap_or_default() as usize;
-    let line_text = &source[line_start.min(source.len())..];
-    let character = match encoding {
-        PositionEncoding::Utf8 => u32::try_from(line_text.len()).unwrap_or(u32::MAX),
-        PositionEncoding::Utf16 => line_text
-            .chars()
-            .map(|character| u32::try_from(character.len_utf16()).unwrap_or_default())
-            .sum(),
-        PositionEncoding::Utf32 => u32::try_from(line_text.chars().count()).unwrap_or(u32::MAX),
-    };
-
-    ls_types::Position::new(line, character)
-}
-
-#[cfg(test)]
-mod tests {
-    use djls_source::LineIndex;
-    use djls_source::PositionEncoding;
-    use tower_lsp_server::ls_types;
-
-    use super::document_end_position;
-
-    #[test]
-    fn document_end_position_handles_trailing_newline() {
-        let source = "first\nsecond\n";
-        let line_index = LineIndex::from(source);
-
-        assert_eq!(
-            document_end_position(source, &line_index, PositionEncoding::Utf16),
-            ls_types::Position::new(2, 0),
-        );
-    }
-
-    #[test]
-    fn document_end_position_uses_client_encoding() {
-        let source = "emoji: 🐍";
-        let line_index = LineIndex::from(source);
-
-        assert_eq!(
-            document_end_position(source, &line_index, PositionEncoding::Utf8),
-            ls_types::Position::new(0, 11),
-        );
-        assert_eq!(
-            document_end_position(source, &line_index, PositionEncoding::Utf16),
-            ls_types::Position::new(0, 9),
-        );
-        assert_eq!(
-            document_end_position(source, &line_index, PositionEncoding::Utf32),
-            ls_types::Position::new(0, 8),
-        );
     }
 }
